Move applying deductions into the class.
